+++ src/main/java/org/owntracks/android/services/ServiceProxy.java
@@ -23,14 +23,6 @@ import de.greenrobot.event.EventBus;
 public class ServiceProxy extends ServiceBindable {
 	private static final String TAG = "ServiceProxy";
 
-	public static final String INTENT_ACTION_PUBLISH_LASTKNOWN = "org.owntracks.android.intent.PUB_LASTKNOWN";
-    public static final String INTENT_ACTION_PUBLISH_LASTKNOWN_MANUAL = "org.owntracks.android.intent.PUB_LASTKNOWN_MANUAL";
-    public static final String INTENT_ACTION_CANCEL_TRANSITION_NOTIFICATION = "org.owntracks.android.intent.INTENT_ACTION_CANCEL_TRANSITION_NOTIFICATION";
-
-    public static final String INTENT_ACTION_PUBLISH_PING = "org.owntracks.android.intent.PUB_PING";
-    public static final String INTENT_ACTION_LOCATION_CHANGED = "org.owntracks.android.intent.LOCATION_CHANGED";
-    public static final String INTENT_ACTION_FENCE_TRANSITION = "org.owntracks.android.intent.FENCE_TRANSITION";
-    public static final String INTENT_ACTION_RECONNECT = "org.owntracks.android.intent.RECONNECT";
 
     public static final String WAKELOCK_TAG_BROKER_PING = "org.owntracks.android.wakelock.broker.ping";
     public static final String WAKELOCK_TAG_BROKER_NETWORK = "org.owntracks.android.wakelock.broker.network";